+++ language/analyze.v1.js
@@ -241,6 +241,79 @@ function analyzeSyntaxInFile (bucketName, fileName) {
   // [END language_syntax_file]
 }
 
+function analyzeEntitySentimentOfText (text) {
+  // [START language_entity_sentiment_string]
+  // Imports the Google Cloud client library
+  const language = require('@google-cloud/language').v1beta2();
+
+  // The text to analyze, e.g. "Hello, world!"
+  // const text = 'Hello, world!';
+
+  // Configure a request containing a string
+  const request = {
+    document: {
+      type: 'PLAIN_TEXT',
+      content: text
+    }
+  };
+
+  // Detects sentiment of entities in the document
+  language.analyzeEntitySentiment(request)
+    .then((results) => {
+      const entities = results[0].entities;
+
+      console.log(`Entities and sentiments:`);
+      entities.forEach((entity) => {
+        console.log(`  Name: ${entity.name}`);
+        console.log(`  Type: ${entity.type}`);
+        console.log(`  Score: ${entity.sentiment.score}`);
+        console.log(`  Magnitude: ${entity.sentiment.magnitude}`);
+      });
+    })
+    .catch((err) => {
+      console.error('ERROR:', err);
+    });
+  // [END language_entity_sentiment_string]
+}
+
+function analyzeEntitySentimentInFile (bucketName, fileName) {
+  // [START language_entity_sentiment_file]
+  // Imports the Google Cloud client library
+  const language = require('@google-cloud/language').v1beta2();
+
+  // The name of the bucket where the file resides, e.g. "my-bucket"
+  // const bucketName = 'my-bucket';
+
+  // The name of the file to analyze, e.g. "file.txt"
+  // const fileName = 'file.txt';
+
+  // Configure a request containing a string
+  const request = {
+    document: {
+      type: 'PLAIN_TEXT',
+      gcsContentUri: `gs://${bucketName}/${fileName}`
+    }
+  };
+
+  // Detects sentiment of entities in the document
+  language.analyzeEntitySentiment(request)
+    .then((results) => {
+      const entities = results[0].entities;
+
+      console.log(`Entities and sentiments:`);
+      entities.forEach((entity) => {
+        console.log(`  Name: ${entity.name}`);
+        console.log(`  Type: ${entity.type}`);
+        console.log(`  Score: ${entity.sentiment.score}`);
+        console.log(`  Magnitude: ${entity.sentiment.magnitude}`);
+      });
+    })
+    .catch((err) => {
+      console.error('ERROR:', err);
+    });
+  // [END language_entity_sentiment_file]
+}
+
 require(`yargs`) // eslint-disable-line
   .demand(1)
   .command(
@@ -279,12 +352,26 @@ require(`yargs`) // eslint-disable-line
     {},
     (opts) => analyzeSyntaxInFile(opts.bucketName, opts.fileName)
   )
+  .command(
+    `entity-sentiment-text <text>`,
+    `Detects sentiment of the entities in a string.`,
+    {},
+    (opts) => analyzeEntitySentimentOfText(opts.text)
+  )
+  .command(
+    `entity-sentiment-file <bucketName> <fileName>`,
+    `Detects sentiment of the entities in a file in Google Cloud Storage.`,
+    {},
+    (opts) => analyzeEntitySentimentInFile(opts.bucketName, opts.fileName)
+  )
   .example(`node $0 sentiment-text "President Obama is speaking at the White House."`)
   .example(`node $0 sentiment-file my-bucket file.txt`, `Detects sentiment in gs://my-bucket/file.txt`)
   .example(`node $0 entities-text "President Obama is speaking at the White House."`)
   .example(`node $0 entities-file my-bucket file.txt`, `Detects entities in gs://my-bucket/file.txt`)
   .example(`node $0 syntax-text "President Obama is speaking at the White House."`)
   .example(`node $0 syntax-file my-bucket file.txt`, `Detects syntax in gs://my-bucket/file.txt`)
+  .example(`node $0 entity-sentiment-text "President Obama is speaking at the White House."`)
+  .example(`node $0 entity-sentiment-file my-bucket file.txt`, `Detects sentiment of entities in gs://my-bucket/file.txt`)
   .wrap(120)
   .recommendCommands()
   .epilogue(`For more information, see https://cloud.google.com/natural-language/docs`)
